==> notebooks/2-window_event_extraction.ipynb <==
import pandas as pd
import numpy as np
import time

from glob import glob
from tqdm import tqdm_notebook as tqdm

import matplotlib.pyplot as plt

from IPython.core.display import display, HTML
display(HTML("<style>.container { width:90% !important; }</style>"))


df_201908 = pd.read_csv("201908_window_events.csv", sep=";", usecols=["Timestamp","Index","Measured LD","Expected LD","Error (FT)","Error (%)","Vehicle ID","Comment"])
df_201909 = pd.read_csv("201909_window_events.csv", sep=";", usecols=["Timestamp","Index","Measured LD","Expected LD","Error (FT)","Error (%)","Vehicle ID","Comment"])
df_201910 = pd.read_csv("201910_window_events.csv", sep=";", usecols=["Timestamp","Index","Measured LD","Expected LD","Error (FT)","Error (%)","Vehicle ID","Comment"])
df_201911 = pd.read_csv("201911_window_events.csv", sep=";", usecols=["Timestamp","Index","Measured LD","Expected LD","Error (FT)","Error (%)","Vehicle ID","Comment"])
df_201912 = pd.read_csv("201912_window_events.csv", sep=";", usecols=["TimeStamp","Record Number","Measured LD (ft)","Expected LD (ft)","Error (ft)","Relative Error (%)","VehicleID","Comment"])
df_202001 = pd.read_csv("202001_window_events.csv", sep=";", usecols=["TimeStamp","Record Number","Measured LD (ft)","Expected LD (ft)","Error (ft)","Relative Error (%)","VehicleID","Comment"])
df_202002 = pd.read_csv("202002_window_events.csv", sep=";", usecols=["TimeStamp","Record Number","Measured LD (ft)","Expected LD (ft)","Error (ft)","Relative Error (%)","VehicleID","Comment"])
df_202003 = pd.read_csv("202003_window_events.csv", sep=";", usecols=["TimeStamp","Record Number","Measured LD (ft)","Expected LD (ft)","Error (ft)","Relative Error (%)","VehicleID","Comment"])

df_201908.columns = ["Timestamp","Record Number", "Measured LD (ft)", "Expected LD (ft)", "Error (ft)", "Relative Error (%)", "VehicleID", "Comment"]
df_201909.columns = ["Timestamp","Record Number", "Measured LD (ft)", "Expected LD (ft)", "Error (ft)", "Relative Error (%)", "VehicleID", "Comment"]
df_201910.columns = ["Timestamp","Record Number", "Measured LD (ft)", "Expected LD (ft)", "Error (ft)", "Relative Error (%)", "VehicleID", "Comment"]
df_201911.columns = ["Timestamp","Record Number", "Measured LD (ft)", "Expected LD (ft)", "Error (ft)", "Relative Error (%)", "VehicleID", "Comment"]
df_201912.columns = ["Timestamp","Record Number", "VehicleID", "Expected LD (ft)", "Measured LD (ft)", "Error (ft)", "Relative Error (%)",  "Comment"]
df_202001.columns = ["Timestamp","Record Number", "VehicleID", "Expected LD (ft)", "Measured LD (ft)", "Error (ft)", "Relative Error (%)",  "Comment"]
df_202002.columns = ["Timestamp","Record Number", "VehicleID", "Expected LD (ft)", "Measured LD (ft)", "Error (ft)", "Relative Error (%)",  "Comment"]
df_202003.columns = ["Timestamp","Record Number", "VehicleID", "Expected LD (ft)", "Measured LD (ft)", "Error (ft)", "Relative Error (%)",  "Comment"]

dfs = [df_201908, df_201909, df_201910, df_201911, df_201912, df_202001, df_202002, df_202003]
df = pd.concat(dfs, sort=True)

df.to_csv("window_events.csv", index=False)
